# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rooms/{id}/availability": {
      "get": {
        "operationId": "RoomsController_availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      }
    },
...
    "/api/v1/reservations": {
      "post": {
        "operationId": "ReservationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      },
      "get": {
        "operationId": "ReservationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/reservations/{id}": {
      "delete": {
        "operationId": "ReservationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reservations"
        ]
      }
    }
  },
  "info": {
    "title": "CSMJU Lab Booking API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateReservationDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}